Condition on length of landmarks_pred, return 0 tensor if true

+++ operation.py
@@ -123,7 +123,7 @@ def generate_landmarks(image):
                     landmarks.append(landmark.z)
 
     if len(landmarks) == 0:
-        return tf.zeros([468, 3])
+        return tf.zeros([468, 3], dtype=tf.float16)
 
     return tf.reshape(landmarks, [468, 3])
